refactor: LottoOrder 클래스를 dto 패키지로 분리
