+++ src/main/java/lotto/dto/LottoOrder.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package lotto.dto;
 
 import java.util.Objects;
