+++ src/test/java/lotto/dto/LottoOrderTest.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package lotto.dto;
 
 import static org.assertj.core.api.Assertions.*;
